+++ client/state/selectors/get-reader-current-stream.js
@@ -1,0 +1,7 @@
+/** @format */
+
+function getCurrentStream( state ) {
+	return state.ui.reader.currentStream;
+}
+
+export default getCurrentStream;
