+++ client/state/selectors/get-reader-stream-next-item.js
@@ -1,0 +1,16 @@
+/** @format */
+
+/**
+ * External dependencies
+ */
+
+/**
+ * Internal dependencie
+ */
+import { getReaderStreamOffsetItem } from './';
+
+function getNextItem( state, currentItem ) {
+	return getReaderStreamOffsetItem( state, currentItem, 1 );
+}
+
+export default getNextItem;
