+++ client/state/selectors/get-reader-stream-prev-item.js
@@ -1,0 +1,16 @@
+/** @format */
+
+/**
+ * External dependencies
+ */
+
+/**
+ * Internal dependencie
+ */
+import { getReaderStreamOffsetItem } from 'state/selectors';
+
+function getPrevItem( state, currentItem ) {
+	return getReaderStreamOffsetItem( state, currentItem, -1 );
+}
+
+export default getPrevItem;
