+++ client/state/selectors/get-reader-stream-transformed-items.js
@@ -1,0 +1,42 @@
+/** @format */
+
+/**
+ * Internal dependencies
+ */
+import { getReaderStream, getReaderFollows } from 'state/selectors';
+import treeSelect from 'lib/tree-select';
+import { injectRecommendations, getDistanceBetweenRecs, combineCards } from 'reader/stream/utils';
+
+/*
+ * getTransformedStreamItems performs the transformations from raw state to data suitable for
+ * Reader cards. That means injecting recs and combining cards.
+ * Signature is:
+ * function( state, { streamKey: string, recsStreamKey: string, shouldCombine: boolean }): Array
+ */
+export const getTransformedStreamItems = treeSelect(
+	( state, { streamKey, recsStreamKey } ) => [
+		getReaderStream( state, streamKey ).items,
+		getReaderStream( state, recsStreamKey ).items,
+		getReaderFollows( state ),
+	],
+	( [ items, recs, follows ], { shouldCombine } ) => {
+		if ( items.length === 0 ) {
+			return [];
+		}
+
+		if ( recs.length > 0 ) {
+			items = injectRecommendations( items, recs, getDistanceBetweenRecs( follows.length ) );
+		}
+
+		if ( shouldCombine ) {
+			items = combineCards( items );
+		}
+		return items;
+	},
+	{
+		getCacheKey: ( { streamKey, recsStreamKey, shouldCombine } ) =>
+			`${ streamKey }${ recsStreamKey }${ shouldCombine }`,
+	}
+);
+
+export default getTransformedStreamItems;
